+++ ML-Models/linear_regression.cpp
@@ -5,37 +5,37 @@
 
 class LinearRegression {
     public:
-    std::vector<float> parameters;
-    float loss;
+    std::vector<double> parameters;
+    double loss;
 
     LinearRegression() {
     }
 
-    std::vector<float> f(std::vector<std::vector<float>> featuresMatrix, std::vector<float> parameters) {
-        std::vector<std::vector<float>> matrix_with_ones = featuresMatrix;
+    std::vector<double> f(std::vector<std::vector<double>> featuresMatrix, std::vector<double> parameters) {
+        std::vector<std::vector<double>> matrix_with_ones = featuresMatrix;
         addOnesToFront(matrix_with_ones);
-        std::vector<float> predictions;
+        std::vector<double> predictions;
 
         for (int i = 0; i < matrix_with_ones.size(); i++) {
-            float dot_product = innerProduct(matrix_with_ones[i], parameters);
+            double dot_product = innerProduct(matrix_with_ones[i], parameters);
             predictions.push_back(dot_product);
         }
         return predictions;
     }
     
-    float calculateLoss(std::vector<std::vector<float>> featuresMatrix, std::vector<float> parameters, std::vector<float> labels) {
-        std::vector<float> predictions = f(featuresMatrix, parameters);
-        float MSE = calculateMSE(predictions, labels);
+    double calculateLoss(std::vector<std::vector<double>> featuresMatrix, std::vector<double> parameters, std::vector<double> labels) {
+        std::vector<double> predictions = f(featuresMatrix, parameters);
+        double MSE = calculateMSE(predictions, labels);
         return MSE;
     }
 
-    void fit(std::vector<std::vector<float>> featuresMatrix, std::vector<float> labels) {
-        std::vector<std::vector<float>> matrix_with_ones = featuresMatrix;
+    void fit(std::vector<std::vector<double>> featuresMatrix, std::vector<double> labels) {
+        std::vector<std::vector<double>> matrix_with_ones = featuresMatrix;
         addOnesToFront(matrix_with_ones);
-        std::vector<std::vector<float>> featuresMatrix_T = takeTranspose(matrix_with_ones);
+        std::vector<std::vector<double>> featuresMatrix_T = takeTranspose(matrix_with_ones);
 
-        std::vector<std::vector<float>> A_T_A = matrixMultiply(featuresMatrix_T, matrix_with_ones);
-        std::vector<float> A_T_b = matrixToVector(matrixMultiply(featuresMatrix_T, vectorToMatrix(labels)));
+        std::vector<std::vector<double>> A_T_A = matrixMultiply(featuresMatrix_T, matrix_with_ones);
+        std::vector<double> A_T_b = matrixToVector(matrixMultiply(featuresMatrix_T, vectorToMatrix(labels)));
 
         this->parameters = solveSystem(A_T_A, A_T_b);
 
@@ -44,7 +44,7 @@ class LinearRegression {
 
     }
 
-    std::vector<float> getPredictions(std::vector<std::vector<float>> featuresMatrix) {
+    std::vector<double> getPredictions(std::vector<std::vector<double>> featuresMatrix) {
         return f(featuresMatrix, parameters);
     }
 };
