+++ libs/Data-Processing/Output.h
@@ -3,8 +3,8 @@
 
 #include <vector>
 
-void printVector(const std::vector<float>& vec);
-void printMatrix(const std::vector<std::vector<float>>& matrix);
+void printVector(const std::vector<double>& vec);
+void printMatrix(const std::vector<std::vector<double>>& matrix);
 void print(std::string toPrint);
 
 #endif
